GHDataCurator/requirements.txt
flask
pandas
pygithub